Added message to stroke failure panic
First, thanks so much for all of your hard work on this library, I truly appreciate your careful implementation, design and documentation.

I've bumped into this panic a few times in the past few weeks and I suspect there are better ways to handle this case, but in the meantime, thought it might be good to at least improve the error message.

For the record, I haven't taken the time to get a local dev environment up and running, or to run any tests. I mainly thought I'd send this for your consideration before I forget about it.

Thanks!